Design, Prod et Mailer puis finito

+++ src/components/Project/GetAllProject.js
@@ -1,8 +1,10 @@
 import React, { useState, useEffect } from 'react'; //React pour JSX
 import { Link, useNavigate } from "react-router-dom";
+import Swal from "sweetalert2";
 import ProjectDataService from "../../services/project.service";
 import AuthService from "../../services/auth.service";
-import App from "../../App"
+import App from "../../App";
+import ToastDisplayer from "../ToastDisplayer";
  
 export default function GetAllProject(props) {
 
@@ -28,14 +30,36 @@ export default function GetAllProject(props) {
   }
 
   function deleteProject(id){ 
+
     //SWAL 
-    ProjectDataService.delete(id)      
-      .then(response => {
-        setProjects(projects.filter(projects => projects.project_id !== id)); //enleve le project supprimé du state pour que ca refresh
-      })
-      .catch(e => {
-        console.log(e);
-      });
+    Swal.fire({
+      position: 'center',
+      icon: 'warning',
+      title: 'Voulez-vous vraiment supprimer ce projet ?',
+      showCancelButton: true,
+      showConfirmButton: true,
+      confirmButtonText: 'Oui',
+      cancelButtonText: "Non",
+      iconColor : 'rgb(232,16,41)',
+      confirmButtonColor: "rgb(232,16,41)",
+      cancelButtonColor: "rgb(61,61,61)",
+      background: 'rgb(33,33,33)',
+      color : 'rgb(255,255,255)'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        //DELETE PROJECT
+        ProjectDataService.delete(id)      
+        .then(response => {
+          setProjects(projects.filter(projects => projects.project_id !== id)); //enleve le project supprimé du state pour que ca refresh
+          ToastDisplayer(false,"Project deleted !");
+        })
+        .catch(e => {
+          ToastDisplayer(true,"An error occured ...");
+          console.log(e);
+        });
+          
+      }
+    });
   }
 
   function redirectToUpdate(id){
